requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
pydantic==2.5.0
numpy==1.24.3
sentence-transformers==2.2.2
python-docx==0.8.11
PyPDF2==3.0.1
beautifulsoup4==4.12.2
markdown==3.5.1
chardet==5.2.0
qdrant-client==1.10.1
httpx==0.26.0
huggingface_hub==0.19.4
